feat: split find into mixin class
