+++ jina/types/arrays/document.py
@@ -1,6 +1,4 @@
-import re
 import json
-import operator
 import warnings
 from abc import abstractmethod
 from collections.abc import MutableSequence, Iterable as Itr
@@ -16,10 +14,10 @@ from typing import (
     Generator,
     BinaryIO,
     TypeVar,
-    Dict,
 )
 
 from .traversable import TraversableSequence
+from .mixins import FilterMixin
 from ..document import Document
 from ...helper import typename, cached_property, cache_invalidate
 from ...proto import jina_pb2
@@ -103,7 +101,7 @@ class DocumentArrayGetAttrMixin:
 
 
 class DocumentArray(
-    TraversableSequence, MutableSequence, DocumentArrayGetAttrMixin, Itr
+    TraversableSequence, MutableSequence, DocumentArrayGetAttrMixin, Itr, FilterMixin
 ):
     """
     :class:`DocumentArray` is a mutable sequence of :class:`Document`.
@@ -442,43 +440,3 @@ class DocumentArray(
             dap.ParseFromString(fp.read())
             da = DocumentArray(dap.docs)
             return da
-
-    def fuzzy_filter(
-        self, regexes: List[Dict[str]], traversal_paths: Iterable[str] = ['r']
-    ):
-        """Filter document array by tags attribute."""
-        filtered = DocumentArray()
-        docs = self.traverse_flat(traversal_paths)
-        for tag_name, regex in regexes:
-            pattern = re.compile(regex)
-            for doc in docs:
-                if re.match(pattern, doc.get(tag_name, '')):
-                    filtered.append(doc)
-        return filtered
-
-    def hard_filter(
-        self,
-        conditions: List[tuple(str, str, str)],
-        traversal_paths: Iterable[str] = ['r'],
-    ):
-        """Filter documents by hard match on tags."""
-        filtered = DocumentArray()
-        docs = self.traverse_flat(traversal_paths)
-        operators_map = {
-            'gt': operator.gt,
-            'lt': operator.lt,
-            'eq': operator.eq,
-            'ne': operator.ne,
-            'ge': operator.ge,
-            'ne': operator.ne,
-        }
-        for condition in conditions:
-            operator_instance = operators_map.get(condition[1], None)
-            if operator_instance:
-                for doc in docs:
-                    tag_name = doc.get(condition[0], None)
-                    tag_value = doc.get(condition[2], None)
-                    if tag_name and tag_value:
-                        if operator_instance(tag_name, tag_value):
-                            filtered.append(doc)
-        return filtered
